+++ .travis/install.sh
@@ -14,9 +14,9 @@ echo "installing miniconda"
 rm -rf $HOME/miniconda
 mkdir -p $HOME/download
 if [ "${TOXENV}" = py27 ]; then
-    curl -s https://repo.continuum.io/miniconda/Miniconda2-latest-${CONDA_INST}-x86_64.sh -o $HOME/download/miniconda.sh
+    curl -s https://repo.anaconda.com/miniconda/Miniconda2-latest-${CONDA_INST}-x86_64.sh -o $HOME/download/miniconda.sh
 else
-    curl -s https://repo.continuum.io/miniconda/Miniconda3-latest-${CONDA_INST}-x86_64.sh -o $HOME/download/miniconda.sh
+    curl -s https://repo.anaconda.com/miniconda/Miniconda3-latest-${CONDA_INST}-x86_64.sh -o $HOME/download/miniconda.sh
 fi
 bash $HOME/download/miniconda.sh -b -p $HOME/miniconda
 
@@ -42,3 +42,8 @@ py36)
     cython "camb>=1" isitgr
   ;;
 esac;
+
+# we have to activate the cond env before we install this
+source activate test-environment
+pip install https://github.com/JoeMcEwen/FAST-PT/archive/v3.0.2.tar.gz --no-deps
+conda deactivate