+++ include/ccl_background.h
@@ -329,6 +329,33 @@ double ccl_mu_MG(ccl_cosmology * cosmo, double a, int *status);
 void ccl_cosmology_compute_distances(ccl_cosmology * cosmo,int *status);
 
 /**
+ * Store user input arrays in splines for the comoving radial distance chi(a), hubble parameter E(a) as well as a(chi).
+ * @param cosmo Cosmological parameters
+ * @param na integer indicating size of array a
+ * @param a scale factor at locations where the input arrays are pre-computed
+ * @param chi_a comoving distance computed at values of a
+ * @param E_a Hubble parameter dividied by H_0 as a function of a
+ * @param status Status flag. 0 if there are no errors, nonzero otherwise.
+ * For specific cases see documentation for ccl_error.c
+ * @return void
+ */
+void ccl_cosmology_distances_from_input(ccl_cosmology * cosmo, int na, double a[], double chi_a[], double E_a[], int *status);
+
+/**
+ * Store user input arrays in splines for the growth factor, growth rate and normalization.
+ * @param cosmo Cosmological parameters
+ * @param na integer indicating size of array a
+ * @param a scale factor at locations where the input arrays are pre-computed
+ * @param growth_arr Growth factor array, defined as D(a)=P(k,a)/P(k,a=1), assuming no scale dependence.
+ * It is assumed that D(a<<1)~a so that D(1.0) will be used for normalization.
+ * @param fgrowth_arr Derivative of the growth array.
+ * @param status Status flag. 0 if there are no errors, nonzero otherwise.
+ * For specific cases see documentation for ccl_error.c
+ * @return void
+ */
+void ccl_cosmology_growth_from_input(ccl_cosmology* cosmo, int na, double a[], double growth_arr[], double fgrowth_arr[], int* status);
+
+/**
  * Compute the growth function and a spline to be stored
  * in the cosmology structure.
  * @param cosmo Cosmological parameters